This is a comprehensive overhaul of the timezone library, python3, adding new functionality, and improving data processing capabilities, fixing issues with negative minuets offsets.
Build System Improvements
- Updated Makefile with cross-platform support (Windows/Unix)
- Added update-tzdata target for automated IANA tzdata updates
- Introduced fitzones target for SensorWatch-specific processing(maybe?? need to track down the valid characters for the bottom row)
- Enhanced clean target to remove all generated files
- Added proper environment variable exports

Data Processing & Generation
- Added automated IANA tzdata download capability (compile_tzdb.py)
- Improved zone name sanitization and formatting for embedded systems
- Enhanced zone packing logic for sensorwatch requirements
- Updated timezone data generation to handle Etc/UTC explicitly
- Added support for %z format specifiers in new tzdb

New Features and Tools
- Added comprehensive unit tests (test_timezone.c) with progress reporting
- Introduced PowerShell and bash provisioning scripts for setup/cleanup
- Updated examples/utility.c for shared utility functions
- Added post_process_sensorwatch.py for zone name sanitization

Codebase Refactoring
- Fixed Python 3 compatibility issues (updated imports, renamed deprecated methods)
- Added OFFSET_INCREMENT = 15 constant
- Updated zone definition structures to support smaller data types
- Improved error handling in timezone calculations
- Refactored zone and rule packing logic

Maintenance
- Added proper encoding handling for all file operations
- Enhanced data validation and error detection

+++ utz.h
@@ -92,7 +92,7 @@ typedef struct uzone_packed_t {
   int8_t offset_inc_minutes;
   uint8_t rules_idx;
   uint8_t rules_len;
-  uint16_t abrev_formatter;
+  uint8_t  abrev_formatter;
 } uzone_packed_t;
 
 /** @struct urule_packed_t
